+++ src/algorithms/games/mod.rs
@@ -6,5 +6,6 @@
 
 pub mod barabasi;
 pub mod erdos_renyi;
+pub mod grg;
 pub mod growing_random;
 pub mod tree;